refactor(core): rename Separators to SubtitleFormat and document joinNonNull
